requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
curl_cffi==0.13.0
dnspython==2.8.0
ecdsa==0.19.2
email-validator==2.3.0
fastapi==0.135.3
frozendict==2.4.7
greenlet==3.3.2
h11==0.16.0
idna==3.11
multitasking==0.0.12
numpy==2.4.4
pandas==3.0.2
passlib==1.7.4
peewee==4.0.3
platformdirs==4.9.4
protobuf==7.34.1
psycopg2-binary==2.9.11
pyasn1==0.6.3
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
pytz==2026.1.post1
redis==7.4.0
requests==2.33.1
rsa==4.9.1
six==1.17.0
soupsieve==2.8.3
SQLAlchemy==2.0.48
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
websockets==16.0
yfinance==1.2.0
